# lambdas/swagger.json
{"openapi":"3.0.0","info":{"title":"URL Shortener API","version":"0.0.1"},"paths":{"/urls/":{"post":{"summary":"Create a short URL","requestBody":{"description":"The original URL and an **optional** custom alias. The custom alias must be unique and have a length between 3 and 20. If none is provided, a random one is generated with length of 8.","content":{"application/json":{"schema":{"type":"object","required":["original_url"],"properties":{"original_url":{"type":"string"},"alias":{"type":"string"}}}}}},"responses":{"201":{"description":"Short URL created successfully","content":{"application/json":{"example":{"alias":"generatedAlias","original_url":"https://www.example.com"}}}},"400":{"description":"Missing request body","content":{"application/json":{"example":{"message":"missing request body"}}}},"409":{"description":"Conflict","content":{"application/json":{"example":{"message":"alias already exists"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"examples":{"emptyOriginalUrl":{"summary":"Empty original_url","value":{"message":"original_url must be provided"}},"invalidOriginalUrlType":{"summary":"Invalid original_url type","value":{"message":"original_url must be a string"}},"invalidAlias":{"summary":"Custom alias length or format invalid","value":{"message":"alias must have length between 3 and 20"}}}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"example":{"message":"something went wrong"}}}}}},"get":{"summary":"List most accessed URLs, ordered by access count","parameters":[{"name":"limit","in":"query","description":"Number of items to return (default is 10). At least 1 and at most 20.","schema":{"type":"integer","minimum":1,"maximum":20}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"example":{"limit":10,"items":[{"alias":"alias1","original_url":"https://www.example.com/1","access_count":10},{"alias":"alias2","original_url":"https://www.example.com/2","access_count":5}]}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"example":{"message":"limit must be at least 1"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"example":{"message":"something went wrong"}}}}}}},"/urls/{alias}":{"get":{"summary":"Redirect by alias","parameters":[{"name":"alias","in":"path","required":true,"description":"The alias for redirection","schema":{"type":"string"}}],"responses":{"302":{"description":"Redirect successful, with Location header","content":{"application/json":{"example":{"message":"Redirecting to https://example.com"}}}},"404":{"description":"Alias not found","content":{"application/json":{"example":{"message":"alias not found"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"example":{"message":"something went wrong"}}}}}}}},"components":{},"tags":[]}